Add in build option to include other code generators.
