+++ CMakeLists.txt
@@ -71,9 +71,6 @@ if (MATERIALX_BUILD_IOS)
 endif()
 
 if (MATERIALX_BUILD_JS)
-    set(MATERIALX_BUILD_GEN_OSL OFF)
-    set(MATERIALX_BUILD_GEN_MSL OFF)
-    set(MATERIALX_BUILD_GEN_MDL OFF)
     set(MATERIALX_BUILD_RENDER OFF)
     set(MATERIALX_BUILD_TESTS OFF)
 endif()
